corrections

+++ 0x00-python_variable_annotations/3-to_str.py
@@ -2,7 +2,7 @@
 """Converts an integer to string"""
 
 
-def to_str(n: int) -> str:
+def to_str(n: float) -> str:
     """Converts an integer to string
 
     Args:
